Added Service and some fixes
Fixed restarting game (before this commit, when finishing game, if you would like to restart, you couldnt do it. Now you can ;) )
