+++ src/com/company/DisplayField.java
@@ -1,10 +1,10 @@
 package com.company;
 
 public class DisplayField {
-    public static Field objectField = TicTacToeGame.objectField;
+     static FieldService fieldService = TicTacToeGame.fieldService;
 
     public static void printField() {
-        int[][] field = objectField.getField(); // ВОТ ТАК НОРМ ???? ИЛИ ВЫЗЫВАТЬ ЧЕРЕЗ САМ КЛАСС ТИПА Field.field ?
+        int[][] field = fieldService.getField(); // ВОТ ТАК НОРМ ???? ИЛИ ВЫЗЫВАТЬ ЧЕРЕЗ САМ КЛАСС ТИПА Field.field ?
         for (int i = 0; i < 50; i++) {
             for (int j = 0; j < 50; j++) {
                 if ((i < 5) | (j < 5) | i > 43 | j > 43)
